The prototype is finished, v0.1 technically
You can add columns and tasks
You can move tasks around each column and inbetween columns
You can change tasks names and colours as well as columns names

This commit also is the start of the big push to convert the current
structure into a fairly strict MVVM version. There are many things this
app is missing and many problems, but nothing more can be built until
the foundation has built solidly, it was rapidly produced to get it to
its current workable form. Now we reduce the technical debt, then
iterate on top and finish it with all the features I want... and a style
upgrade.